Cookie deletion bugfix
The cookie creation utility (auth.set_cookie) was setting path=/ by
default. But the cookie deletion utility was not, so the deletion path
defaulted to the current document path.  This was not a problem for
traditional nb because the current path was always "/" (there are no
subdirectories on nb.mit.edu).  But when nb annotates html the
document path might not be /, which means deleting the cookie at the
current path fails to delete the cookie that was set at /, which
prevented logout from working.