A hack to fix xpath matching in xhtml documents.
The HTML annotator uses xpaths to define the locations of comments.
The xpath is determined and stored when a comment is recorded.
When comments are displayed/highlighted, their location in the
document is determined using document.evaluate(xpath...).   In
HTML documents this works fine.  But in xhtml documents, the
document.evaluate method pays attention to namespaces.  All tags
in the document are assumed, at parse time, to be in the xhtml
namespace defined in the DOCTYPE declaration.  On the other hand,
the xpath for the comment was constructed with no namespace prefix,
so its contents are assumed to be in the *null* namespace so fail
to match the corresponding tags in the document.

As a backup, I've added a small hack that, when the xpath fails to
find any math the "normal" way, tries faking a namespace into the
xpath so that it will match properly in xhtml documents.  This is only
invoked when the normal method fails, so should not affect performance
in the common case.

Long term, instead of using xpaths and xquery which has this fragility
we may want to use CSS selectors, which are more forgiving.