Fix code to handle ?c=<number> parameter in url
NB looks up html annotations based on url.  that means a reference to
an in-page comment (which adds a ?c=<comment-id> parameter at the end
of the url) is a different url so notes aren't found.  Changed code to
ignore ? in determining commented url

This is a temporary fix for ucdavis experiments.  it will break nb on
sites that differentiate pages (that should get different annotations)
by using a query string.