Corrected failing 'NB$' in window tests
We've allowed for override of jquery by checking for presence of a
global NB$ variable.  But recently we've wrapped our code in a
function() {} body for isolation.  But now that NB$ is a *local*
variable, our "in window" test fails. I've replaced those tests with
tests for whether NB$ is defined (as a local variable).