Add ability to directly add users to ensemble
We had issues last quarter with making sure students were
added to all the ensembles we used for a course because they
had to find several emails and accept invitations from all
of them or (later) use a subscription link with which to
add themselves.

This commit adds a checkbox in the invite_users_dialog for
ensemble administrators to directly add students to
ensembles without requiring further action from the students
(after verifying their email address and entering their
name). Emails are still sent if the user does not already
exist, regardless of whether or not the checkbox is checked.